feat: effect deps 추가 및 API 분리
- 마지막 피드가 getMore를 바꾸고 API는 이 변수를 참조하여 effect를 일으킨다
